♻️ refactor folder simpleStaking -> tasks_simple_staking
