+++ tasks/tasks_simple_staking/configureStaking.ts
